+++ components/keyemulate/impl/src/main/java/com/flipperdevices/keyemulate/model/EmulateButtonState.kt
@@ -20,7 +20,10 @@ sealed class EmulateButtonState {
     object ForbiddenFrequencyDialog : Inactive()
 
     @Stable
-    data class Active(val progress: EmulateProgress) : EmulateButtonState()
+    data class Active(
+        val progress: EmulateProgress,
+        val config: EmulateConfig
+    ) : EmulateButtonState()
 }
 
 enum class DisableButtonReason {